original Histogram Dialog merged with Ehistogramdialog
